Contact form!
I'm currently working on the contact form, and specifically trying to figure out what data binding is and how to apply two way data binding to be able to collect data from my form when the user inputs their info. I'm a bit stuck right now because it does not seem to be updating correctly at the moment!